+++ src/pits_mras/training/__init__.py
@@ -6,7 +6,11 @@ are implemented.
 """
 
 from pits_mras.training.cotrain import cotraining_loop
-from pits_mras.training.hinf_minmax import hinf_minmax_train, hji_residual
+from pits_mras.training.hinf_minmax import (
+    hinf_minmax_from_dynamics,
+    hinf_minmax_train,
+    hji_residual,
+)
 from pits_mras.training.irl_trainer import train_irl_critic
 from pits_mras.training.pretrain import pretrain_pitnn
 
@@ -15,5 +19,6 @@ __all__ = [
     "cotraining_loop",
     "train_irl_critic",
     "hinf_minmax_train",
+    "hinf_minmax_from_dynamics",
     "hji_residual",
 ]